refactor(cypress): split commands in their own files
