+++ packages/e2e-cypress/src/helpers/index.ts
@@ -1,3 +1,6 @@
+/// <reference types="cypress" />
+/// <reference types="@cypress/vue" />
+
 export * from './install-quasar-plugin';
 export * from './commands';
